requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
# torch>=2.1.0 --index-url https://download.pytorch.org/whl/cpu
# torchvision>=0.16.0 --index-url https://download.pytorch.org/whl/cpu
transformers>=4.36.0
accelerate>=0.25.0
safetensors
pillow
opencv-python
pydicom
numpy
scikit-image
google-generativeai
